InMemoryObjectMessagesStore - not working yet

+++ modules/direction/src/main/java/edu/agh/iga/adi/giraph/direction/io/data/IgaMessageWritable.java
@@ -30,7 +30,7 @@ public class IgaMessageWritable implements WritableComparable {
   @Override
   public int compareTo(Object o) {
     IgaMessageWritable other = (IgaMessageWritable) o;
-    return (int) (igaMessage.getSrcId() - other.igaMessage.getSrcId());
+    return igaMessage.getSrcId() - other.igaMessage.getSrcId();
   }
 
   public IgaMessage getIgaMessage() {
